+++ Bonjour/MYBonjourService.m
@@ -181,8 +181,8 @@ NSString* const kBonjourServiceResolvedAddressesNotification = @"BonjourServiceR
                           port: (uint16_t)port
                      txtRecord: (NSData*)txtData
 {
-    LogTo(Bonjour, @"%@: hostname=%@, port=%u, txt=%u bytes", 
-          self, hostname, port, txtData.length);
+    LogTo(Bonjour, @"%@: hostname=%@, port=%u, txt=%lu bytes", 
+          self, hostname, port, (unsigned long)txtData.length);
 
     if (port!=_port || !$equal(hostname,_hostname)) {
         self.hostname = hostname;
