+++ TCP/TCPListener.m
@@ -153,7 +153,7 @@ static CFSocketRef closeSocket( CFSocketRef socket ) {
         _ipv4socket = [self _openProtocol: PF_INET address: (struct sockaddr*)&addr4 error: &error];
         if( ! _ipv4socket ) {
             if( error.code==EADDRINUSE && _pickAvailablePort && _port<0xFFFF ) {
-                LogTo(TCPVerbose,@"%@: port busy, trying %hu...",self,_port+1);
+                LogTo(TCPVerbose,@"%@: port busy, trying %u...",self,_port+1);
                 self.port += 1;        // try the next port
             } else {
                 if( outError ) *outError = error;
@@ -334,7 +334,7 @@ static void TCPListenerAcceptCallBack(CFSocketRef socket, CFSocketCallBackType t
         if( _bonjourTXTRecord ) {
             data = [NSNetService dataFromTXTRecordDictionary: _bonjourTXTRecord];
             if( data )
-                LogTo(TCP,@"%@: Set %u-byte TXT record", self,data.length);
+                LogTo(TCP,@"%@: Set %lu-byte TXT record", self,(unsigned long)data.length);
             else
                 Warn(@"TCPListener: Couldn't convert txt dict to data: %@",_bonjourTXTRecord);
         } else
@@ -353,7 +353,7 @@ static void TCPListenerAcceptCallBack(CFSocketRef socket, CFSocketCallBackType t
 - (void)netService:(NSNetService *)sender didNotPublish:(NSDictionary *)errorDict
 {
     self.bonjourError = [[errorDict objectForKey:NSNetServicesErrorCode] intValue];
-    LogTo(TCP,@"%@: Failed to advertise %@: error %i",self,sender,self.bonjourError);
+    LogTo(TCP,@"%@: Failed to advertise %@: error %li",self,sender,(long)self.bonjourError);
     [_netService release];
     _netService = nil;
 }
